refactor(persistence): Refactor package structure
