+++ persistence/model/src/main/java/org/company/persistence/model/HasId.java
@@ -1,4 +1,4 @@
-package model;
+package org.company.persistence.model;
 
 public interface HasId {
     Integer getId();
