+++ persistence/repository/src/main/java/org/company/persistence/repository/ProductRepository.java
@@ -1,15 +1,15 @@
-package repository;
+package org.company.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import model.Product;
+import org.company.persistence.model.Product;
 
-@RepositoryRestResource(path = "products")
+@Repository
 public interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
     @Modifying
     @Transactional
